Moved middleware to src
