+++ src/middleware/swagger-security.js
