+++ src/middleware/swagger-ui/favicon-16x16.png
