+++ src/middleware/swagger-ui/favicon-32x32.png
