+++ src/middleware/swagger-ui/oauth2-redirect.html
